ai-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.3.0
ultralytics==8.3.73
numpy==2.1.3
